Make resouceloader a standalone package
