+++ buildSrc/src/main/kotlin/tech/antibytes/gradle/util/test/config/ConfigBase.kt
@@ -1,0 +1,11 @@
+/*
+ * Copyright (c) 2022 Matthias Geisler (bitPogo) / All rights reserved.
+ *
+ * Use of this source code is governed by Apache v2.0
+ */
+
+package tech.antibytes.gradle.util.test.config
+
+abstract class ConfigBase {
+    val group = "tech.antibytes.test-utils-kmp"
+}
