+++ settings.gradle.kts
@@ -17,7 +17,8 @@ include(
     ":test-utils",
     ":test-utils-annotations",
     ":test-utils-coroutine",
-    ":test-utils-ktor"
+    ":test-utils-ktor",
+    ":test-utils-resourceloader"
 )
 
 buildCache {
