+++ test-utils-annotations/build.gradle.kts
@@ -8,7 +8,7 @@ import tech.antibytes.gradle.configuration.ensureIosDeviceCompatibility
 import tech.antibytes.gradle.configuration.isIdea
 import tech.antibytes.gradle.dependency.Dependency
 import tech.antibytes.gradle.util.test.dependency.Dependency as LocalDependency
-import tech.antibytes.gradle.util.test.config.TestUtilsConfiguration
+import tech.antibytes.gradle.util.test.config.AnnotationsConfiguration
 
 plugins {
     id("org.jetbrains.kotlin.multiplatform")
@@ -21,12 +21,12 @@ plugins {
     id("tech.antibytes.gradle.coverage")
 }
 
-group = TestUtilsConfiguration.group
+group = AnnotationsConfiguration.group
 
 antiBytesPublishing {
-    packageConfiguration = TestUtilsConfiguration.publishing.packageConfiguration
-    repositoryConfiguration = TestUtilsConfiguration.publishing.repositories
-    versioning = TestUtilsConfiguration.publishing.versioning
+    packageConfiguration = AnnotationsConfiguration.publishing.packageConfiguration
+    repositoryConfiguration = AnnotationsConfiguration.publishing.repositories
+    versioning = AnnotationsConfiguration.publishing.versioning
 }
 
 android {
