+++ test-utils-resourceloader/src/androidMain/AndroidManifest.xml
@@ -1,0 +1,7 @@
+<!--
+  ~ Copyright (c) 2022 Matthias Geisler (bitPogo) / All rights reserved.
+  ~
+  ~ Use of this source code is governed by Apache v2.0
+  -->
+
+<manifest/>
