+++ test-utils-resourceloader/src/jsTest/kotlin/tech/antibytes/util/test/resourceloader/IsBrowser.kt
@@ -1,0 +1,11 @@
+/*
+ * Copyright (c) 2022 Matthias Geisler (bitPogo) / All rights reserved.
+ *
+ * Use of this source code is governed by Apache v2.0
+ */
+
+package tech.antibytes.util.test.resourceloader
+
+import kotlinx.browser.window
+
+actual val isBrowser: Boolean = jsTypeOf(window) != "undefined"
