+++ test-utils-resourceloader/src/nativeTest/kotlin/tech/antibytes/util/test/resourceloader/IsBrowser.kt
@@ -1,0 +1,9 @@
+/*
+ * Copyright (c) 2022 Matthias Geisler (bitPogo) / All rights reserved.
+ *
+ * Use of this source code is governed by Apache v2.0
+ */
+
+package tech.antibytes.util.test.resourceloader
+
+actual val isBrowser: Boolean = false
