--- test-utils/src/commonTest/kotlin/tech/antibytes/util/test/IsBrowser.kt
@@ -1,9 +1,0 @@
-/*
- * Copyright (c) 2022 Matthias Geisler (bitPogo) / All rights reserved.
- *
- * Use of this source code is governed by Apache v2.0
- */
-
-package tech.antibytes.util.test
-
-expect val isBrowser: Boolean
